+++ aws/openvpn/instance.tf
@@ -8,17 +8,22 @@ resource "random_password" "this" {
 
 /*
 https://registry.terraform.io/providers/hashicorp/tls/latest/docs/resources/private_key
+Ignore `tls_private_key` creation if you want to use existing key `ec2_public_key`.
  */
 resource "tls_private_key" "management_ssh_key" {
-  count     = var.create_management_key_pair ? 1 : 0
+  count     = (var.create_management_key_pair && var.ec2_public_key == null) ? 1 : 0
   algorithm = "RSA"
   rsa_bits  = 4096
 }
 
+/**
+Provides an EC2 key pair resource. A key pair is used to control login access to EC2 instances.
+https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/key_pair
+ */
 resource "aws_key_pair" "management_ssh_key" {
   count      = var.create_management_key_pair ? 1 : 0
-  key_name   = var.vpn_name
-  public_key = tls_private_key.management_ssh_key[0].public_key_openssh
+  key_name   = var.key_name != null ? var.key_name : var.vpn_name
+  public_key = var.ec2_public_key != null ? var.ec2_public_key : tls_private_key.management_ssh_key[0].public_key_openssh
 }
 
 /*
@@ -29,7 +34,7 @@ resource "aws_instance" "this" {
 
   ami           = data.aws_ami.debian.id
   instance_type = var.instance_type
-  key_name      = var.create_management_key_pair ? var.vpn_name : null
+  key_name      = var.create_management_key_pair ? aws_key_pair.management_ssh_key[0].key_name : null
   subnet_id     = var.subnet_id
 
   vpc_security_group_ids = concat(
@@ -48,10 +53,6 @@ resource "aws_instance" "this" {
 
   user_data_replace_on_change = true
   user_data                   = templatefile("${path.module}/scripts/install_openvpn.sh", local.openvpn_config)
-
-  tags = {
-    "Name" = var.vpn_name
-  }
 
   lifecycle {
     ignore_changes = [
@@ -66,7 +67,7 @@ resource "aws_instance" "replacable" {
 
   ami           = data.aws_ami.debian.id
   instance_type = var.instance_type
-  key_name      = var.create_management_key_pair ? var.vpn_name : null
+  key_name      = var.create_management_key_pair ? aws_key_pair.management_ssh_key[0].key_name : null
   subnet_id     = var.subnet_id
 
   vpc_security_group_ids = concat(
@@ -85,10 +86,6 @@ resource "aws_instance" "replacable" {
 
   user_data_replace_on_change = true
   user_data                   = templatefile("${path.module}/scripts/install_openvpn.sh", local.openvpn_config)
-
-  tags = {
-    "Name" = var.vpn_name
-  }
 
   lifecycle {
     ignore_changes = [